Replace in-memory storage with SQLite database
The system has been refactored to replace the in-memory data storage system with a persistent SQLite database. A new entity, MetricData, has been created to correspond to tables in the new database. The Entity Framework Core was used to migrate the previously stored data. All actions related to the old in-memory system have been updated to function with the new SQLite database system.